users/requirements.txt
fastapi[standard]
sqlalchemy
psycopg2-binary
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
autopep8
isort